requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0.12
django-autokit==0.2.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-pg==0.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.4
setuptools==78.1.1
sqlparse==0.5.3
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
wheel==0.45.1
